+++ typing.d.ts
@@ -71,6 +71,12 @@ export interface QNA {
     fileUrl: string; // 첨부파일
     password: string; //비번
     isAnswered: boolean; // 답변유무
+}
+
+export interface Thesis {
+    content: string; // 내용
+    createAt: string; // 생성일
+    fileUrl: string; // 첨부파일
     title: string; // 제목
     writer: string; // 작성자
 }
